requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.20.0
requests==2.32.0
openai==1.52.0
python-dotenv==1.0.0
faiss-cpu==1.9.0.post1
langchain==0.2.17
langchain-community==0.2.17
langchain-openai==0.1.25
langchain-huggingface==0.0.3
pypdf==4.3.1
numpy==1.26.2
pytest==7.4.2
